+++ urgrep-pkg.el
@@ -1,2 +1,2 @@
-;; Generated package description from urgrep.el  -*- no-byte-compile: t -*-
-(define-package "urgrep" "0.5.3snapshot0.20250701.205050" "Universal recursive grep" '((emacs "27.1") (compat "29.1.0.1") (project "0.3.0")) :commit "92bcc1d8e30da62a67287e58c732fe5b9373b22f" :keywords '("grep" "search") :url "https://github.com/jimporter/urgrep")
+;; Generated package description from urgrep.el  -*- mode: lisp-data; no-byte-compile: t -*-
+(define-package "urgrep" "0.6.0snapshot0.20250709.114307" "Universal recursive grep" '((emacs "28.1") (compat "29.1.0.1") (project "0.3.0")) :commit "4dd2a9a60196aa17640593659ee8aa550a9bf2e0" :keywords '("grep" "search") :url "https://github.com/jimporter/urgrep")